requirements.txt
dotenv~=0.9.9
python-dotenv~=1.1.0
flask~=3.1.1
requests~=2.32.4
pycryptodome~=3.23.0
pyjwt~=2.10.1